*feat*: Added smart detection for resetting the color variable.
If the color variable is an object, then it will be reset to it's default. This will allow the object to not have to be recreated.
Otherwise it is set directly to null.

This is very useful and this feature has been added because of the `aParticleGenerator` library. Inside of the `aParticleGenerator` library when transitioning the color a lot and fast the objective is to keep the color object intact to prevent it from needing to recreated. So when the particle needs to be cleaned, it should keep the same object.