Update aRecycle.js
Added some sanity checks. If a Diob is recycled, and it belongs to an interface it is now removed from that interface. 
If an interface is recycled it will no longer have it's mapName variable assigned to null since that variable is useless to an interface, and may be being used by a developer.