Update aRecycle.js
The library now cleans a Diob before it is deleted. This ensures that any binding information or things attached to it will be released. 
This also makes sure that if the Diob was an interface, and that if is a WebBox then the WebBox it uses is removed from the DOM.