Separate server and client logic better
This commit also removes the `pObject` argument inside of `isInCollection`. It was a waste of an argument. Since the type is passed the plugin can infer if it is an object or not. It does not need to be explicitly passed. This frees us the space to use for strictly arguments to be passed into `instance.onDumped`.

Removed circular references from being logged in errors.

Fixed a sanity check from throwing the wrong error.

Removed COLLECTION_LIMIT and instead references it via a variable belonging to the plugin now.

Adds a sanity check before using `VS.Client`

Removed additional checks for typeof()